server/pyproject.toml
[project]
name = "dnd-summarizer-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Xian",email = "michaelkxian@gmail.com"}
]
package-mode = false
requires-python = "==3.12.10"
dependencies = [
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.7,<0.129.0)",
    "pyttsx3 (>=2.99,<3.0)",
    "transformers (==4.57.6)",
    "torch (==2.8.0)",
    "openai-whisper (>=20250625,<20250626)",
    "typer[all] (>=0.23.0,<0.24.0)",
    "accelerate (>=1.12.0,<2.0.0)",
    "safetensors (>=0.7.0,<0.8.0)",
    "peft (>=0.18.1,<0.19.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
